risk type and risk status added as string in api of approvals

+++ Risk_Management_RiskEX_Backend/MappingConfig.cs
@@ -24,7 +24,7 @@ namespace Risk_Management_RiskEX_Backend
                 .ForMember(dest => dest.RiskName, opt => opt.MapFrom(src => src.RiskName))
                 .ForMember(dest => dest.Description, opt => opt.MapFrom(src => src.Description))
                 .ForMember(dest => dest.PlannedActionDate, opt => opt.MapFrom(src => src.PlannedActionDate))
-                .ForMember(dest => dest.RiskType, opt => opt.MapFrom(src => (int)src.RiskType))
+                .ForMember(dest => dest.RiskType, opt => opt.MapFrom(src => src.RiskType))
                 .ForMember(dest => dest.OverallRiskRating, opt => opt.MapFrom(src => src.OverallRiskRating))
                 .ForMember(dest => dest.RiskStatus, opt => opt.MapFrom(src => src.RiskStatus));
